refactor: apply Java coding standard formatting
