+++ src/main/java/command/CommandType.java
@@ -1,3 +1,8 @@
+package command;
+
+/**
+ * Represents the different types of commands that can be executed.
+ */
 public enum CommandType {
     LIST,
     MARK,
